Init apollo implementation
