+++ frontend/package.json
@@ -3,12 +3,15 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@apollo/react-hooks": "^3.1.3",
     "@types/jest": "24.0.21",
     "@types/node": "12.12.5",
     "@types/react": "16.9.11",
     "@types/react-dom": "^16.9.3",
     "@types/react-router-dom": "^5.1.2",
+    "apollo-boost": "^0.4.4",
     "bootstrap": "^4.3.1",
+    "graphql": "^14.5.8",
     "msal": "^1.1.3",
     "node-sass": "^4.13.0",
     "react": "^16.11.0",
