+++ frontend/src/Components/App/App.tsx
@@ -1,29 +1,16 @@
 import React from 'react';
 import Login from '../Login/Login';
 import Dashboard from '../Dashboard/Dashboard';
-import * as Msal from 'msal';
-import { BrowserRouter, Route, Switch, Redirect } from 'react-router-dom';
+import createAuthenticator from '../../Helpers/AuthenticatorFactory'
+import { BrowserRouter, Route, Switch } from 'react-router-dom';
+import ApolloClient from 'apollo-boost';
+import { ApolloProvider } from '@apollo/react-hooks';
 
 const App: React.FC = () => {
 
-  const policy: string = process.env.REACT_APP_B2C_POLICY;
-  const clientId: string = process.env.REACT_APP_B2C_CLIENTID;
-  const tenantName: string = process.env.REACT_APP_B2C_TENANT_NAME;
-
-  let msalConfig: Msal.Configuration = {
-    auth: {
-      clientId: clientId,
-      authority: `https://${tenantName}.b2clogin.com/${tenantName}.onmicrosoft.com/${policy}`,
-      validateAuthority: false,
-      postLogoutRedirectUri: "http://localhost:3000/"
-    },
-    cache: {
-      cacheLocation: "sessionStorage",
-      storeAuthStateInCookie: true
-    }
-  };
-
-  let msalInstance = new Msal.UserAgentApplication(msalConfig);
+  // Dependencies
+  const msalInstance = createAuthenticator();
+  const apolloClient = new ApolloClient();
 
   const logout: React.FC = () => {
     msalInstance.logout();
@@ -32,16 +19,18 @@ const App: React.FC = () => {
 
   return (
     <Login authenticator={msalInstance}>
-      <BrowserRouter>
-        <Switch>
-          <Route exact path="/">
-            <Dashboard />
-          </Route>
-          <Route exact path="/logout">
-            {logout}
-          </Route>
-        </Switch>
-      </BrowserRouter>
+      <ApolloProvider client={apolloClient}>
+        <BrowserRouter>
+          <Switch>
+            <Route exact path="/">
+              <Dashboard />
+            </Route>
+            <Route exact path="/logout">
+              {logout}
+            </Route>
+          </Switch>
+        </BrowserRouter>
+      </ApolloProvider>
     </Login>
   );
 }
